+++ core/src/com/unciv/ui/screens/newgamescreen/GameOptionsTable.kt
@@ -122,6 +122,7 @@ class GameOptionsTable(
         if (gameParameters.isOnlineMultiplayer){
             checkboxTable.addAnyoneCanSpectateCheckbox()
             checkboxTable.addRequireSameVersionCheckbox()
+            checkboxTable.addForbidReloadCheckbox()
             selectBoxTable.addDurationSelectBox("Time until skip turn:", GameParameters::minutesUntilSkipTurn, 1, 0, 0)
             selectBoxTable.addDurationSelectBox("Total time to play:", GameParameters::minutesUntilForceResign, 3, 0, 0)
             selectBoxTable.addDurationSelectBox("Time recovered per turn:", GameParameters::minutesRecoveredPerTurn, 3, 0, 0)
@@ -225,6 +226,10 @@ class GameOptionsTable(
     private fun Table.addRequireSameVersionCheckbox() =
             addCheckbox("Require same game version to join", gameParameters.requireSameVersion)
             { gameParameters.requireSameVersion = it }
+
+    private fun Table.addForbidReloadCheckbox() =
+            addCheckbox("Forbid reloading", gameParameters.forbidReload)
+            { gameParameters.forbidReload = it }
 
     private fun Table.addEnableEspionageCheckbox() =
         addCheckbox("Enable Espionage", gameParameters.espionageEnabled)